refactor: split auth controller from app module to auth module
