+++ server/src/auth/auth.module.ts
@@ -4,6 +4,7 @@ import { PassportModule } from "@nestjs/passport";
 
 import { UsersModule } from "../users/users.module";
 
+import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
 import { jwtConstants } from "./constants";
 import { JwtStrategy } from "./jwt.strategy";
@@ -18,6 +19,7 @@ import { LocalStrategy } from "./local.strategy";
       signOptions: { expiresIn: "1h" },
     }),
   ],
+  controllers: [AuthController],
   providers: [AuthService, LocalStrategy, JwtStrategy],
   exports: [AuthService],
 })
